// src/app/shared/model/scoring-result-h-obj.model.ts
export class ScoringResultHObj{
    VerfResultId: number;
    MrScoringMethodCode: string;
    OfficeCode: string;
    ReqBy: string;
    ReqDate: string;
    ScoringAlias: string;
    ScoringResultHId: number;
    ScoringTrxNo: string;
    ScoringValue: number;
    TrxSourceNo: string;
    TrxSourceType: string;

    constructor() {  
        this.ScoringValue=0;
    }
}